poetry_demo/pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["amitaiturkel <amitai.turkel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tox = "^4.11.3"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
asyncio-nats-client = "^0.11.5"
aiohttp = "^3.8.5"
httpx = "^0.25.0"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "4.1.0"
pytest-asyncio = "0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
